Programs API v2 (#1966)
* Added tests, fixture and utility functions

* oops, put nplus 1 back & reparameterize the fixture

* reformat

* Add to Programs as well

* reformat

* add v2

* updated per Nathan's feedback

* merge + update for feedback

* moved to v2, versioned tests and serializers as well

* remove random copy pasta

* Example of catalog data fixture

* added v1 and v2 programs. having urls issues

* reformat

* namespace fixed. Paring down programs list a little.

* Fixing programs

* reformat

* reformat

* fix merge

* testing

* fix extra urls for records/program

* less calls

* reformat

* v2 serializer test

* refactor v1 structure + add financial page

* move v2 stuff

* missing import

* reformat

* remove empty file

* reformat

* imports

* how though

* more mocks eaten

* add .live()

* update for additional query due to flex form

* reformat

* test failure

* fmt

* fix patch

---------
