+++ cms/serializers_test.py
@@ -11,7 +11,7 @@ from cms.factories import (
     ProgramPageFactory,
 )
 from cms.models import FlexiblePricingRequestForm
-from cms.serializers import CoursePageSerializer
+from cms.serializers import CoursePageSerializer, ProgramPageSerializer
 from courses.factories import (
     CourseFactory,
     ProgramFactory,
@@ -304,3 +304,108 @@ def test_serialized_course_finaid_form_url(
                 related_fa_page.slug
                 in serialized_output["financial_assistance_form_url"]
             )
+
+
+def test_serialize_program_page(
+    mocker, fully_configured_wagtail, staff_user, mock_context
+):
+    fake_image_src = "http://example.com/my.img"
+    patched_get_wagtail_src = mocker.patch(
+        "cms.serializers.get_wagtail_img_src", return_value=fake_image_src
+    )
+
+    program = ProgramFactory(page=None)
+    program_page = ProgramPageFactory(program=program)
+    financial_assistance_form = FlexiblePricingFormFactory(
+        selected_program_id=program.id, parent=program_page
+    )
+    rf = RequestFactory()
+    request = rf.get("/")
+    request.user = staff_user
+
+    data = ProgramPageSerializer(
+        instance=program_page, context=program_page.get_context(request)
+    ).data
+    assert_drf_json_equal(
+        data,
+        {
+            "feature_image_src": fake_image_src,
+            "page_url": program_page.url,
+            "financial_assistance_form_url": f"{program_page.get_url()}{financial_assistance_form.slug}/",
+            "description": bleach.clean(program_page.description, tags=[], strip=True),
+            "live": True,
+            "length": program_page.length,
+            "effort": program_page.effort,
+            "price": None,
+        },
+    )
+
+
+def test_serialize_program_page__with_related_financial_form(
+    mocker, fully_configured_wagtail, staff_user, mock_context
+):
+    fake_image_src = "http://example.com/my.img"
+    patched_get_wagtail_src = mocker.patch(
+        "cms.serializers.get_wagtail_img_src", return_value=fake_image_src
+    )
+
+    program = ProgramFactory(page=None)
+    program_page = ProgramPageFactory(program=program)
+    other_program = ProgramFactory(page=None)
+    other_program_page = ProgramPageFactory(program=other_program)
+    financial_assistance_form = FlexiblePricingFormFactory(
+        selected_program_id=other_program.id, parent=other_program_page
+    )
+    program.add_related_program(other_program)
+    rf = RequestFactory()
+    request = rf.get("/")
+    request.user = staff_user
+
+    data = ProgramPageSerializer(
+        instance=program_page, context=program_page.get_context(request)
+    ).data
+    assert_drf_json_equal(
+        data,
+        {
+            "feature_image_src": fake_image_src,
+            "page_url": program_page.url,
+            "financial_assistance_form_url": f"{program_page.get_url()}{financial_assistance_form.slug}/",
+            "description": bleach.clean(program_page.description, tags=[], strip=True),
+            "live": True,
+            "length": program_page.length,
+            "effort": program_page.effort,
+            "price": None,
+        },
+    )
+
+
+def test_serialize_program_page__no_financial_form(
+    mocker, fully_configured_wagtail, staff_user, mock_context
+):
+    fake_image_src = "http://example.com/my.img"
+    patched_get_wagtail_src = mocker.patch(
+        "cms.serializers.get_wagtail_img_src", return_value=fake_image_src
+    )
+
+    program = ProgramFactory(page=None)
+    program_page = ProgramPageFactory(program=program)
+    rf = RequestFactory()
+    request = rf.get("/")
+    request.user = staff_user
+
+    data = ProgramPageSerializer(
+        instance=program_page, context=program_page.get_context(request)
+    ).data
+    assert_drf_json_equal(
+        data,
+        {
+            "feature_image_src": fake_image_src,
+            "page_url": program_page.url,
+            "financial_assistance_form_url": "",
+            "description": bleach.clean(program_page.description, tags=[], strip=True),
+            "live": True,
+            "length": program_page.length,
+            "effort": program_page.effort,
+            "price": None,
+        },
+    )
