+++ courses/urls/v2/urls.py
@@ -1,0 +1,10 @@
+from django.urls import include, re_path
+from rest_framework import routers
+
+from courses.views import v2
+
+app_name = "courses"
+router = routers.SimpleRouter()
+router.register(r"programs", v2.ProgramViewSet, basename="programs_api")
+
+urlpatterns = router.urls
